example 15: convert the Mona-Lisa parameter side + CPU rasteriser to float
The genome and the host evaluation now use the GFloat parameter hierarchy:

  * GImageIndividual: every genome parameter becomes a GConstrainedFloatObject
    with a GFloatGaussAdaptor (the four sigma args cast to float, ad_prob stays
    double); fitnessCalculation streamlines as std::vector<float>.
  * GMonaLisaProblem: the shared rasteriser (Target, calculateCorners,
    pointInTriangle, score, renderToRGB, scoreAgainstTarget) is float; the target
    image is stored as float.
  * GMonaLisaGPUMarshaller + GImagePOM follow suit (float streamline / hostEvaluate
    / render).

The courtier GPU framework is still double-typed, so the marshaller bridges the
float genome to the double buffers (flatten widens, hostEvaluate narrows, the
problem-constant blob stays double) -- the CUDA kernel remains double and keeps
working unchanged. This makes the pure-CPU path (backend=cpu) genuinely float and
proves the GFloat hierarchy on a real problem; native FP32 on the GPU is the
follow-on pass (templating the GPU framework on a scalar type).

Verified: full build clean, ctest 65/65, backend=cuda converges (787162->709752)
and backend=cpu (float) converges.

+++ geneva/examples/15_GCUDAWorker/GImagePOM.hpp
@@ -245,7 +245,7 @@ private:
                 goa->Interface::GOptimizerIT<oa::GBase>::getBestIterationIndividual<GImageIndividual>();
 
             // Rasterise the candidate genome at the target resolution and write it out.
-            std::vector<double> parVec;
+            std::vector<float> parVec;
             best_ptr->streamline(parVec);
             const MonaLisa::Target &tgt = MonaLisa::target();
             std::vector<unsigned char> rgb;
